Skip flaky test on Windows (#507)
