+++ nevergrad/functions/multiobjective/test_core.py
@@ -3,8 +3,6 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-import nevergrad as ng
-from nevergrad.functions import MultiobjectiveFunction
 from . import core
 
 
@@ -28,11 +26,27 @@ def test_multiobjective_function() -> None:
     assert front == expected_front, f"Expected {expected_front} but got {front}"
 
 
-def test_readme_example() -> None:
-    f = MultiobjectiveFunction(multiobjective_function=lambda x: (x[0]**2 + x[1]**2, (x[0]-1.)**2 + (x[1]-1.)**2), upper_bounds=[2.5, 2.5])
+# pylint: disable=redefined-outer-name,unsubscriptable-object,unused-variable,unused-import
+def test_doc_multiobjective() -> None:
+    # DOC_MULTIOBJ_0
+    import nevergrad as ng
+    from nevergrad.functions import MultiobjectiveFunction
+    import numpy as np
+
+    f = MultiobjectiveFunction(multiobjective_function=lambda x: [np.sum(x**2), np.sum((x - 1)**2)], upper_bounds=[2.5, 2.5])
+    print(f(np.array([1.0, 2.0])))
+
     optimizer = ng.optimizers.CMA(parametrization=3, budget=100)  # 3 is the dimension, 100 is the budget.
-    optimizer.minimize(f)
+    recommendation = optimizer.minimize(f)
+
     # The function embeds its Pareto-front:
+    print("My Pareto front:", [x[0][0] for x in f.pareto_front()])
+
+    # It can also provide a subset:
+    print("My Pareto front:", [x[0][0] for x in f.pareto_front(2, subset="random")])
+    print("My Pareto front:", [x[0][0] for x in f.pareto_front(2, subset="loss-covering")])
+    print("My Pareto front:", [x[0][0] for x in f.pareto_front(2, subset="domain-covering")])
+    # DOC_MULTIOBJ_1
     assert len(f.pareto_front()) > 1
     assert len(f.pareto_front(2, "loss-covering")) == 2
     assert len(f.pareto_front(2, "domain-covering")) == 2
